% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/condense_spaces.R
\name{condense_spaces}
\alias{condense_spaces}
\title{Trim and remove double spaces from the input strings}
\usage{
condense_spaces(x)
}
\arguments{
\item{x}{A vector}
}
\value{
A vector with spaces removed
}
\description{
Trim and remove double spaces from the input strings
}
\details{
Beginning, double and ending
         spaces are also removed from the strings.
}
\examples{
\dontrun{x2 <- condense_spaces(x)}
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("sorvi")
}
\keyword{utilities}

