% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_municipality_codes.R
\name{convert_municipality_codes}
\alias{convert_municipality_codes}
\title{Munucipality code conversions}
\usage{
convert_municipality_codes(ids = NULL, municipalities = NULL)
}
\arguments{
\item{ids}{NULL}

\item{municipalities}{NULL}
}
\value{
Depending on the input. Converted id or name vector, or full conversion table.
}
\description{
Conversions between municipality codes and names
}
\examples{
 \dontrun{conversion.table <- convert_municipality_codes()}
}
\references{
See citation("sorvi")
}
\author{
Leo Lahti \email{louhos@googlegroups.com}
}
\keyword{utilities}
