% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_municipality_names.R
\name{convert_municipality_names}
\alias{convert_municipality_names}
\title{Convert municipality names into standard versions harmonized within the package}
\usage{
convert_municipality_names(municipality.names)
}
\arguments{
\item{municipality.names}{municipality names to convert}
}
\value{
standardized municipality names
}
\description{
Convert municipality names into standard versions harmonized within the package
}
\examples{
\dontrun{tmp <- convert_municipality_names("Koski.Tl")}
}
\references{
See citation("sorvi")
}
\author{
Leo Lahti \email{louhos@googlegroups.com}
}
\keyword{internal}
