\name{ConvertMMLToRData}
\alias{ConvertMMLToRData}
\title{Convert MML shape objects into RData format. For detailed example, see https://github.com/louhos/sorvi/wiki/Maanmittauslaitos}
\usage{
  ConvertMMLToRData(MML, output.data.dir = "RData/")
}
\arguments{
  \item{MML}{output from GetShapeMML(input.data.dir = ".")

  Returns:}
}
\value{
  output data directory name
}
\description{
  Arguments:
}
\examples{
#
}
\author{
  Leo Lahti \email{louhos@googlegroups.com}
}
\references{
  See citation("sorvi")
}
\keyword{utilities}

