\name{dfsort}
\alias{dfsort}
\title{Sort data frame}
\usage{
  dfsort(df, sortvar, ...)
}
\arguments{
  \item{df}{data.frame to be sorted by the specified
  columns}

  \item{sortvar}{variable/s according which the data.frame
  shall be sorted}

  \item{...}{Other arguments to pass}
}
\value{
  data.frame (sorted version)
}
\description{
  Sort data frame
}
\examples{
# dfsort(df, x, -z)
}
\author{
  Leo Lahti \email{louhos@googlegroups.com}
}
\references{
  See citation("sorvi")
}
\keyword{utilities}

