% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fi.en.maakunnat}
\alias{fi.en.maakunnat}
\title{fi.en.maakunnat data documentation}
\format{list}
\usage{
#translations <- load_sorvi_data("translations")
}
\description{
Mappings between Finnish and English province (maakunta) names
}
\author{
Leo Lahti \email{louhos@googlegroups.com}
}
\keyword{data}
