% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_statfi_classifications.R
\name{get_classification_df}
\alias{get_classification_df}
\title{Get classifications from Statistics Finland API}
\usage{
get_classification_df(
  name = NULL,
  date = NULL,
  localId = NULL,
  lang = "fi",
  content = "data",
  meta = "max"
)
}
\arguments{
\item{name}{name of the classification, often in format "NAME_1"}

\item{date}{date of the classification, in format `"1999-01-01"` (year, month, day)}

\item{localId}{localId of the object. This field supersedes the name and date fields, if not null}

\item{lang}{either "fi" (default), "en", "sv" or "all"}

\item{content}{can only be "data" (default)}

\item{meta}{can be either "min" or "max" (default)}
}
\description{
Downloads classifications from Statistics Finland
}
\details{
Concatenated together the name and date parameters form the localId of the
classification item. If only localId is used as a parameter, it supersedes the separate
date and name parameters.

Available classifications and their localId's can be viewed in
\url{https://data.stat.fi/api/classifications/v2/classifications?content=url}
or
\url{https://data.stat.fi/api/classifications/v2/classifications?content=data&lang=fi}.
}
\examples{
\donttest{
maakunta <- get_classification_df("maakunta_1", "20200101", lang = "fi")
maakunta_all <- get_classification_df("maakunta_1", "20200101", lang = "all")
}

}
