% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gender_fi.R
\name{get_gender_fi}
\alias{get_gender_fi}
\title{Finnish name-gender mappings}
\usage{
get_gender_fi(download = FALSE)
}
\arguments{
\item{download}{If TRUE, the data downloaded from the original
source. Otherwise, a readily preprocessed version from the sorvi R
package is used.}
}
\value{
A data_frame with the field "name" (first name), followed
by the population frequencies and population counts for that name.
}
\description{
Finnish name-gender mappings from the Finnish
population register (VRK 3/2016).
}
\details{
Information from the Finnish Population register. All
first names for living Finnish citizens that live in Finland and
abroad. Only names with frequency n>10 are included.
 Source: avoindata.fi service and Vaestorekisterikeskus (VRK).
 URL: https://www.avoindata.fi/data/fi/dataset/none
 Version: 3/2016
 Data license CC-BY 4.0
The data was downloaded with this function as sorvi::get_gender_table_fi()
and integrated in the sorvi package. The frequencies of male/female names
are added after the download as some names occur for both genders.
}
\examples{
gender_FI <- get_gender_fi()
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See cite("sorvi")
}
\keyword{data}

