% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ifpi_charts.R
\name{get_ifpi_charts}
\alias{get_ifpi_charts}
\title{Get CRAN download statistics}
\usage{
get_ifpi_charts(channel = "radio", year = 2020, week = 50)
}
\arguments{
\item{channel}{Options: "radio", "albumit", "singlet", "fyysiset-albumit"}

\item{year}{year as numeric, default is 2020}

\item{week}{week as numeric, default is 50. Please note that number of weeks 
differ between years. For simplicity's sake valid weeks are set to be 
between 1 and 53. Use e.g. `lubridate::isoweek` to check how many weeks a
given year has.}
}
\value{
tibble
}
\description{
Download chart position data from ifpi.fi
}
\details{
Web scraping function that is inspired by Sauravkaushik8 Kaushik's
blog post "Beginner's Guide on Web Scraping in R" on analyticsvidhya.com. 
Downloads chart data from Musiikkituottajat - IFPI Finland ry website.
Please note that this function works only with IFPI Finland website!
}
\seealso{
Original tutorial in \url{https://www.analyticsvidhya.com/blog/2017/03/beginners-guide-on-web-scraping-in-r-using-rvest-with-hands-on-knowledge/}
}
\author{
Pyry Kantanen <pyry.kantanen@gmail.com>
}
