% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_municipalities_by_year.R
\name{get_municipalities}
\alias{get_municipalities}
\title{Select Municipalities by Year}
\source{
Data attribution: FinnONTO Consortium: \url{https://seco.cs.aalto.fi/projects/finnonto/}
}
\usage{
get_municipalities(year = 2002, type = "data.frame")
}
\arguments{
\item{year}{a year between 1865-2020}

\item{type}{either "data.frame" or "sf"}
}
\value{
a data.frame or sf object
}
\description{
From a larger dataset containing historical municipalities, 
pick a certain year and return an output that contains the most recent 
information on each municipality.
}
\details{
See dataset "kunnat1865_2021"
}
\author{
Pyry Kantanen
}
