% Generated by roxygen2 (4.0.0): do not edit by hand
\name{get_municipality_info_mml}
\alias{get_municipality_info_mml}
\title{Get information of Finnish municipalities from Land Survey Finland 2013.
(C) Maanmittauslaitos MML 2013. For details of MML data, see
help(GetShapeMML).}
\usage{
get_municipality_info_mml()
}
\value{
A data frame with municipality data
}
\description{
Get information of Finnish municipalities from Land Survey Finland 2013.
(C) Maanmittauslaitos MML 2013. For details of MML data, see
help(GetShapeMML).
}
\examples{
\dontrun{tab <- get_municipality_info_mml()}
}
\author{
Leo Lahti \email{louhos@googlegroups.com}
}
\references{
See citation("sorvi")
}
\keyword{utilities}

