% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_municipality_info_statfi}
\alias{get_municipality_info_statfi}
\title{Get information of Finnish municipalities from Statistics Finland 2013}
\usage{
get_municipality_info_statfi(verbose = TRUE, ...)
}
\arguments{
\item{verbose}{verbose}

\item{...}{Arguments to be passed}
}
\value{
A data frame with municipality data
}
\description{
Get information of Finnish municipalities from Statistics Finland 2013
}
\examples{
\dontrun{df <- get_municipality_info_statfi()}
}
\author{
Leo Lahti \email{louhos@googlegroups.com}
}
\references{
See citation("sorvi")
}
\keyword{utilities}

