% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_population_register}
\alias{get_population_register}
\title{Retrieve population register data}
\usage{

  get_population_register(url = "http://vrk.fi/default.aspx?docid=5127&site=3&id=0")
}
\arguments{
\item{url}{String specifying the URL containing the population register data.}
}
\value{
data.frame with municipality information.
}
\description{
Retrieves municipality data from population register.
The url should be changed when there are updates:
}
\examples{
\dontrun{df <- get_population_register()}
}
\author{
Leo Lahti \email{louhos@googlegroups.com}
}
\references{
See citation("sorvi")
}
\keyword{utilities}

