% Generated by roxygen2 (4.0.0): do not edit by hand
\name{get_province_info_wikipedia}
\alias{get_province_info_wikipedia}
\title{Get information of Finnish provinces.}
\usage{
get_province_info_wikipedia(...)
}
\arguments{
\item{...}{Arguments to be passed}
}
\value{
A data frame. With the following entries: Maakunta: province; Pinta-ala: area; Vakiluku: population; Vaestotiheys: population density
}
\description{
Get information of Finnish provinces.
}
\examples{
\dontrun{tab <- get_province_info_wikipedia()}
}
\author{
Leo Lahti \email{louhos@googlegroups.com}
}
\references{
See citation("sorvi")
}
\keyword{utilities}

