\name{GetAllAanestykset}
\alias{GetAllAanestykset}
\title{Hakee kaikki 채채nestykset eduskuntarajapinnasta}
\usage{
  GetAllAanestykset()
}
\arguments{
  \item{no}{params}
}
\value{
  list
}
\description{
  Hakee kaikki 채채nestykset eduskuntarajapinnasta
}
\examples{
# aanestykset <- GetAllAanestykset()
}
\author{
  Juuso Haapanen
  \email{sorvi-commits@lists.r-forge.r-project.org}
}
\references{
  See citation("sorvi")
}
\keyword{eduskunta}

