\name{GetEdustajaData}
\alias{GetEdustajaData}
\title{Hakee edustajan tiedot äänestystunnisteen perusteella}
\usage{
  aanestykset <- GetEdustajaData(tunniste)
}
\value{
  Äänestyksen tulos data.framena, jossa kansanedustajan nimi, puolue ja äänestystulos
}
\description{
  Hakee tunnisteella äänestykset edustajaittain.
}
\author{
  Juuso Haapanen
  \email{sorvi-commits@lists.r-forge.r-project.org}
}

