\name{GetElectedCandidates}
\alias{GetElectedCandidates}
\title{GetElectedCandidates}
\usage{
  GetElectedCandidates(year, election, election.district)
}
\arguments{
  \item{year}{election year}

  \item{election}{eletion type (municipal / parliament /
  president / ...)}
}
\value{
  data.frame
}
\description{
  Get data on elected candidates
}
\examples{
#
}
\author{
  Leo Lahti \email{louhos@googlegroups.com}
}
\references{
  See citation("sorvi")
}
\keyword{utilities}

