\name{GetMember}
\alias{GetMember}
\title{GetMember
Get members of a particular council}
\arguments{
  \item{x}{object}

  \item{name}{name of a person}
}
\value{
  candidate object
}
\description{
  GetMember Get members of a particular council
}
\author{
  Joona Lehtomaki \email{louhos@googlegroups.com}
}

