\name{GetOikotie}
\alias{GetOikotie}
\title{Preprocess Oikotie data}
\usage{
  GetOikotie(data.dir = ".")
}
\arguments{
  \item{data.dir}{data directory for unzipped
  http://www2.hs.fi/extrat/hsnext/oikotie-data.zip}
}
\value{
  list with Oikotie myynnit data
}
\description{
  Preprocess data about Finnish apartment prices from
  Oikotie in years 2010-2011. First download and unzip data
  from http://www2.hs.fi/extrat/hsnext/oikotie-data.zip,
  and store the resulting CSV files.
}
\details{
  Arguments:
}
\author{
  Juuso Parkkinen \email{louhos@googlegroups.com}
}

