\name{GetParties}
\alias{GetParties}
\title{GetParties
Get Parties from Council class object}
\arguments{
  \item{x}{Council class object

  Returns:}
}
\value{
  parties object
}
\description{
  Arguments:
}
\examples{
#
}
\author{
  Joona Lehtomaki \email{louhos@googlegroups.com}
}
\references{
  See citation("sorvi")
}
\keyword{utilities}

