% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_issue_api.R
\name{github_issue_stats}
\alias{github_issue_stats}
\title{GitHub issues statistics}
\usage{
github_issue_stats(owner = "ropengov", repo = "geofi")
}
\arguments{
\item{owner}{Repository owner / organization. Default is "ropengov"}

\item{repo}{Repository name. Default is "geofi"}
}
\value{
tibble
}
\description{
Get statistics about GitHub issues from GitHub API.
}
\seealso{
GitHub Issues API documentation: 
\url{https://docs.github.com/en/rest/reference/issues}

Original "analyze GitHub stuff with R" repository:
\url{https://github.com/jennybc/analyze-github-stuff-with-r}
}
\author{
Original scripts by Jennifer Bryan (jennybc), function by 
Pyry Kantanen <pyry.kantanen@gmail.com>
}
