\name{haeHakusanalla}
\alias{haeHakusanalla}
\title{Hakee hakusanalla äänestyksiä eduskuntarajapinnasta}
\usage{
  haeHakusanalla(hakusana)
}
\arguments{
  \item{hakusana}{string}
}
\value{
  data.frame
}
\description{
  Hakee hakusanalla äänestyksiä eduskuntarajapinnasta
}
\examples{
# luonto <- haeHakusanalla('luonto')
}
\author{
  Juuso Haapanen
  \email{sorvi-commits@lists.r-forge.r-project.org}
}
\references{
  See citation("sorvi")
}
\keyword{eduskunta}

