% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_names.R
\name{harmonize_names}
\alias{harmonize_names}
\title{Harmonize names}
\usage{
harmonize_names(x, synonymes, remove.unknown = FALSE, mode = "exact.match",
  verbose = FALSE)
}
\arguments{
\item{x}{A character vector}

\item{synonymes}{synonyme table with the fields 'synonyme' and 'name'}

\item{remove.unknown}{Logical. Remove terms that do not have synonymes.}

\item{mode}{'exact.match' replaces the terms based on the synonyme list if an exact match is  found; 'match' replaces the parts that match synonymes; 'recursive' replaces all (sub)strings recursively in the same order as in the synonyme table}

\item{verbose}{verbose}
}
\value{
Vector of harmonized terms
}
\description{
Harmonize names
}
\details{
Map the input vector to harmonized versions based on the synonyme table.
}
\examples{
\dontrun{x2 <- harmonize_names(x, synonymes)}
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("sorvi")
}
\keyword{utilities}

