% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_url.R
\name{is_url}
\alias{is_url}
\title{is_url}
\usage{
is_url(s)
}
\arguments{
\item{s}{input object to check}
}
\value{
TRUE/FALSE indicating whether the input string is a valid URL.
}
\description{
Check if the given object is an url string
}
\examples{
is_url("http://aa.px")
}
\references{
See citation("sorvi")
}
\author{
Leo Lahti \email{louhos@googlegroups.com}
}
\keyword{utilities}
