% Generated by roxygen2 (4.0.0): do not edit by hand
\name{korvaa_skandit}
\alias{korvaa_skandit}
\title{Replace special characters with standard ones.}
\usage{
korvaa_skandit(s)
}
\arguments{
\item{s}{string from which the special chars should be removed}
}
\value{
string with special chars replaced by standard ones
}
\description{
Replace special characters with standard ones.
}
\note{
iconv function provides better tools for these purposes and is now the main tool
This function is kept for compatibility with the older versions.
}
\examples{
korvaa_skandit("my.string.here") # if no, special chars, the same string is returned
}
\author{
Leo Lahti \email{louhos@googlegroups.com}
}
\references{
See citation("sorvi")
}
\keyword{utilities}

