% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_sorvi_data.R
\name{load_sorvi_data}
\alias{load_sorvi_data}
\title{Supporting Data}
\usage{
load_sorvi_data(data.id = "translation_provinces")
}
\arguments{
\item{data.id}{data ID to download (see details)}
}
\value{
a data.frame
}
\description{
Load custom data sets.
}
\details{
This function is a thin wrapper to `get_classification_df()`. The only
value data.id parameter will accept as argument is `translation_provinces`,
which will return a data.frame object containing the English and Finnish
names of Finnish provinces from 2025 classification.
}
\examples{
translations <- load_sorvi_data("translation_provinces")
}
\references{
See citation("sorvi")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
