\name{LoadData}
\alias{LoadData}
\title{LoadData}
\usage{
  LoadData(data.id,
    url = "http://beta.datavaalit.fi/storage/louhos/",
    verbose = TRUE)
}
\arguments{
  \item{data.id}{data ID to download (suffix before .rda).
  Investigate the contents of the url path to check
  data.ids}

  \item{url}{data url}

  \item{verbose}{verbose

  Return:}
}
\value{
  translations
}
\description{
  Arguments:
}
\examples{
# translations <- LoadData("translations")
}
\author{
  Leo Lahti \email{louhos@googlegroups.com}
}
\references{
  See citation("sorvi")
}
\keyword{utilities}

