% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{polygons1909_2009}
\alias{polygons1909_2009}
\title{Municipality geometries}
\format{
A simple feature with 860 rows and 3 variables:
\describe{
  \item{x}{Universal Resource Identifier (URI) for each municipality instance in time. For example: http://www.yso.fi/onto/sapo/Maalahti(1908-1972)}
  \item{kunta_nro}{Municipality code, a unique 3-digit code (001-999) assigned for each municipality that stays the same as long as the municipality exists. For example: "475"}
  \item{geometry}{A single list column with geometries}
}
}
\source{
Original data downloaded from ONKI.fi website on 04 Aug 2022: \url{http://onki.fi/en/browser/overview/sapo}
Data attribution: FinnONTO Consortium: \url{https://seco.cs.aalto.fi/projects/finnonto/}
}
\usage{
polygons1909_2009
}
\description{
A simple feature containing the URIs, municipality codes and geometries of 
municipalities in time. The starting point and end point of each municipality
can be determined by combining polygons1909_2009 with another dataset that
contains such information.
}
\keyword{datasets}
