% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickdf.R
\name{quickdf}
\alias{quickdf}
\title{Quick Data Frame}
\usage{
quickdf(x)
}
\arguments{
\item{x}{named list of equal length vectors}
}
\value{
data.frame
}
\description{
Speedups in data.frame creation.
}
\details{
This is 20x faster than as.data.frame.
}
\examples{
 \dontrun{quickdf(x)}
}
\author{
HW/LL
}
\references{
Idea borrowed from \url{http://adv-r.had.co.nz/Profiling.html}.
}
\keyword{utilities}

