% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_synonymes.R
\name{read_synonymes}
\alias{read_synonymes}
\title{Read synonyme table}
\usage{
read_synonymes(file, mode = "list", sep = ";", self.match = FALSE,
  include.lowercase = FALSE, ignore.empty = FALSE, sort = FALSE,
  verbose = FALSE, remove.ambiguous = TRUE)
}
\arguments{
\item{file}{Input file}

\item{mode}{The input file type: "list" or "table"; see details.}

\item{sep}{Separator mark.}

\item{self.match}{Include self matches}

\item{include.lowercase}{Include lowercase}

\item{ignore.empty}{Ignore the empty cases}

\item{sort}{Sort synonymes}

\item{verbose}{verbose}

\item{remove.ambiguous}{Remove ambiguous terms.}
}
\value{
Synonyme data frame with the fields 'name' (the selected term) and 'synonyme' (the alternative terms).
}
\description{
Read synonyme table.
}
\details{
If mode = "list", each row of the input file corresponds to a unique entry with potentially multiple name variants, separated by semicolon. The first element gives the selected version of the name, the subsequent elements list synonymes that will be mapped to the selected version. If mode = "table", the file has two columns where each row corresponds to a unique entry and has the selected name and a single alternative name.
}
\examples{
\dontrun{syn <- read_synonymes(file)}
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("sorvi")
}
\keyword{utilities}

