\name{ReadDatavaalit}
\alias{ReadDatavaalit}
\title{Load data sets from datavaalit.fi web service}
\usage{
  ReadDatavaalit(data.id)
}
\arguments{
  \item{data.id}{Data set ID}
}
\value{
  rjson object
}
\description{
  Load data sets from datavaalit.fi web service
}
\author{
  Leo Lahti
  \email{sorvi-commits@lists.r-forge.r-project.org}
}

