% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ropengov_storage_path.R
\name{ropengov_storage_path}
\alias{ropengov_storage_path}
\title{ropengov_storage_path}
\usage{
ropengov_storage_path(x)
}
\arguments{
\item{x}{Characted string key defining which storage path
is requested.

Return:}
}
\value{
URL for Louhos data
}
\description{
ropengov_storage_path
}
\examples{
url <- ropengov_storage_path("mml")

}
\references{
See citation("sorvi")
}
\author{
Leo Lahti \email{louhos@googlegroups.com}
}
\keyword{utilities}
