% Generated by roxygen2 (4.0.0): do not edit by hand
\name{set_breaks}
\alias{set_breaks}
\title{Set breaks for color palette. Internal function.}
\usage{
set_breaks(mat, interval = 0.1)
}
\arguments{
\item{mat}{data matrix or vector for which the breaks will be deterined}

\item{interval}{interval of color breaks}
}
\value{
A vector of breakpoints
}
\description{
Set breaks for color palette. Internal function.
}
\author{
Leo Lahti \email{louhos@googlegroups.com}
}
\references{
See citation("sorvi")
}
\keyword{internal}

