% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sorvi-package.R
\docType{package}
\name{sorvi-package}
\alias{sorvi}
\alias{sorvi-package}
\title{Algorithmic Tools for Open Data in Finland}
\description{
Brief summary of the sorvi package
}
\details{
\tabular{ll}{
Package: \tab sorvi\cr
Type: \tab Package\cr
Version: \tab See sessionInfo() or DESCRIPTION file\cr
Date: \tab 2010-2015\cr
License: \tab BSD-2-clause\cr
LazyLoad: \tab yes\cr
}

Open Data tools for Finland
}
\examples{
library(sorvi)
}
\author{
Leo Lahti, Juuso Parkkinen, Jussi Paananen, Joona Lehtomaki, Einari Happonen, and Juuso Haapanen \email{louhos@googlegroups.com}
}
\references{
See citation("sorvi")
\url{https://github.com/rOpenGov/sorvi}
}
\keyword{package}

