\name{SorviData}
\alias{SorviData}
\title{SorviData}
\usage{
  SorviData(data.id,
    sorvi.data.url = "http://beta.datavaalit.fi/storage/louhos/")
}
\arguments{
  \item{data.id}{}

  \item{sorvi.data.url}{Data URL path}
}
\value{
  url connection
}
\description{
  SorviData
}
\examples{
# load(SorviData("MML"))
}
\author{
  Leo Lahti
  \email{sorvi-commits@lists.r-forge.r-project.org}
}
\references{
  See citation("sorvi")
}
\keyword{utilities}

