\name{SotkanetRegions}
\alias{SotkanetRegions}
\title{Description:
SotkanetRegions retrieves Sotkanet regions data from
http://www.sotkanet.fi/rest/1.1/regions}
\usage{
  SotkanetRegions(type = "table")
}
\arguments{
  \item{type}{Return format ("table" or "raw")

  Returns:}
}
\value{
  sotkanet json query in selected format
}
\description{
  Arguments:
}
\examples{
#
}
\author{
  Einari Happonen / Opasnet / Louhos. Maintainer:
  Louhos/Opasnet \email{louhos@googlegroups.com}
}
\references{
  See citation("sorvi")
}
\keyword{utilities}

