% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\name{strip}
\alias{strip}
\title{strip string i.e. remove spaces from the beginning and end}
\usage{
strip(s)
}
\arguments{
\item{s}{string or character vector}
}
\value{
stripped string
}
\description{
strip string i.e. remove spaces from the beginning and end
}
\examples{
strip(c("s ", " a"))
}
\author{
Leo Lahti \email{louhos@googlegroups.com}
}
\references{
See citation("sorvi")
}
\keyword{utilities}

