% Generated by roxygen2 (4.0.0): do not edit by hand
\name{strstrip}
\alias{strstrip}
\title{Remove spaces from a string (single string or vector/list of strings).}
\usage{
strstrip(s)
}
\arguments{
\item{s}{string or vector/list of strings}
}
\value{
string without spaces
}
\description{
Remove spaces from a string (single string or vector/list of strings).
}
\examples{
strstrip("a b") # returns "ab"
}
\author{
Leo Lahti \email{louhos@googlegroups.com}
}
\references{
See citation("sorvi")
}
\keyword{utilities}

