% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wfs.R
\name{TurkuWFS}
\alias{TurkuWFS}
\title{Turku-specific WFS client}
\description{
A specialized subclass of \code{WFSClient} that defaults to the City of Turku’s
geoservice WFS endpoint.
}
\examples{
\dontrun{
  turku <- TurkuWFS$new()
  layers <- turku$list_features()
  sf_obj <- turku$get_feature("GIS:Aanestysalueet", crs = "EPSG:3067")
}

}
\section{Super class}{
\code{\link[sorvi:WFSClient]{sorvi::WFSClient}} -> \code{TurkuWFS}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TurkuWFS-new}{\code{TurkuWFS$new()}}
\item \href{#method-TurkuWFS-clone}{\code{TurkuWFS$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="sorvi" data-topic="WFSClient" data-id="get_feature"><a href='../../sorvi/html/WFSClient.html#method-WFSClient-get_feature'><code>sorvi::WFSClient$get_feature()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sorvi" data-topic="WFSClient" data-id="list_features"><a href='../../sorvi/html/WFSClient.html#method-WFSClient-list_features'><code>sorvi::WFSClient$list_features()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TurkuWFS-new"></a>}}
\if{latex}{\out{\hypertarget{method-TurkuWFS-new}{}}}
\subsection{Method \code{new()}}{
Create a new Turku WFS client
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TurkuWFS$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TurkuWFS-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TurkuWFS-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TurkuWFS$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
