% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synonymes.R
\name{check_synonymes}
\alias{check_synonymes}
\title{check_synonymes}
\usage{
check_synonymes(synonymes, include.lowercase = TRUE, verbose = FALSE,
  sort = FALSE, self = FALSE)
}
\arguments{
\item{synonymes}{synonymes data.frame with the self-explanatory fields 'name' and 'synonyme'.}

\item{include.lowercase}{Include lowercase versions of the synonymes}

\item{verbose}{verbose}

\item{sort}{Sort synonymes}

\item{self}{Ensure that each name is synonyme for itself (this may
cause ambiguous mappings, use with care !)}
}
\value{
Polished synonyme table
}
\description{
Check synonyme table.
}
\details{
Remove duplicated information. Ensure identical matches
are included in synonyme list.
}
\examples{
\dontrun{s <- check_synonymes(synonymes)}
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("sorvi")
}
\keyword{utilities}

