% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/condense_spaces.R
\name{condense_spaces}
\alias{condense_spaces}
\title{condense_spaces}
\usage{
condense_spaces(x)
}
\arguments{
\item{x}{A vector}
}
\value{
A polished vector
}
\description{
Trim spaces
}
\examples{
\dontrun{x2 <- condense_spaces(x)}
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

