% Generated by roxygen2 (4.0.1): do not edit by hand
\name{convert_municipality_codes}
\alias{convert_municipality_codes}
\title{Conversions between municipality codes and names}
\usage{
convert_municipality_codes(ids = NULL, municipalities = NULL)
}
\arguments{
\item{ids}{NULL}

\item{municipalities}{NULL}
}
\value{
Depending on the input. Converted id or name vector, or full conversion table.
}
\description{
Conversions between municipality codes and names
}
\examples{
\dontrun{conversion.table <- convert_municipality_codes()}
}
\author{
Leo Lahti \email{louhos@googlegroups.com}
}
\references{
See citation("sorvi")
}
\keyword{utilities}

