% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran_downloads.R
\name{cran_downloads}
\alias{cran_downloads}
\title{Get CRAN download statistics}
\usage{
cran_downloads(
  pkgs = "all",
  output = "tibble",
  sum = "by_month",
  plot.scale = 11,
  use.cache = TRUE
)
}
\arguments{
\item{pkgs}{Package name(s)}

\item{output}{"tibble" or "plot"}

\item{sum}{"by_month", "by_year" or "total"}

\item{plot.scale}{integer, default is 10. Smaller numbers decrease the size of
plot elements, larger numbers make them larger.}

\item{use.cache}{Cache downloaded statistics. Default is TRUE}
}
\value{
sf object
}
\description{
Produces a tibble or a visualization of package download
   statistics.
}
\examples{
\dontrun{
df <- cran_downloads(pkgs = "eurostat", sum = "total", use.cache = FALSE)
kable(df)

## Compare two packages
p1 <- cran_downloads(pkgs = "eurostat", sum = "by_year", output = "plot")
p2 <- cran_downloads(pkgs = "osmar", sum = "by_year", output = "plot")
gridExtra::grid.arrange(p1, p2, nrow = 2)
}

}
\author{
Leo Lahti, Pyry Kantanen <pyry.kantanen@gmail.com>
}
