% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_municipality_info_mml.R
\name{get_municipality_info_mml}
\alias{get_municipality_info_mml}
\title{Municipality Info}
\usage{
get_municipality_info_mml(...)
}
\arguments{
\item{...}{Arguments to be passed}
}
\value{
A data frame with municipality data
}
\description{
Get information of Finnish municipalities from
Land Survey Finland. (C) Maanmittauslaitos MML 2013.
}
\examples{
\dontrun{tab <- get_municipality_info_mml()}
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("sorvi")
}
\keyword{utilities}

