% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_postal_code_info}
\alias{get.postal.codes}
\alias{get_postal_code_info}
\title{Get Finnish postal codes vs. municipalities table from Wikipedia.}
\usage{
get_postal_code_info(...)
}
\arguments{
\item{...}{Arguments to be passed}
}
\value{
A data frame with following fields: postal.code: postal code; municipality: Name of the municipality (kunnan nimi); municipality.scandless: Municpality name without special chars
}
\description{
Get Finnish postal codes vs. municipalities table from Wikipedia.
}
\examples{
\dontrun{postal.code.table <- get_postal_code_info()}
}
\author{
Juuso Parkkinen and Leo Lahti \email{louhos@googlegroups.com}
}
\references{
See citation("sorvi")
}
\keyword{utilities}

