\name{GetAllAanestykset}
\alias{GetAllAanestykset}
\title{Hakee eduskunnan äänestysrajanpinnan kautta kaikki äänestykset}
\usage{
  aanestykset <- GetAllAanestykset()
}
\value{
  lista kaikista äänestyksistä
}
\description{
  Hakee kaikki äänestykset
}
\author{
  Juuso Haapanen
  \email{sorvi-commits@lists.r-forge.r-project.org}
}

