\name{GetAllAanestykset}
\alias{GetAllAanestykset}
\title{Hakee kaikki aanestykset eduskuntarajapinnasta}
\usage{
  GetAllAanestykset()
}
\arguments{
  \item{no}{params}
}
\value{
  list
}
\description{
  Hakee kaikki aanestykset eduskuntarajapinnasta
}
\examples{
# aanestykset <- GetAllAanestykset()
}
\author{
  Juuso Haapanen
  \email{sorvi-commits@lists.r-forge.r-project.org}
}
\references{
  citation("sorvi")
}
\keyword{eduskunta}

