\name{GetAllAanestykset}
\alias{GetAllAanestykset}
\title{Hakee kaikki aanestykset eduskuntarajapinnasta}
\usage{
  GetAllAanestykset(...)
}
\arguments{
  \item{...}{parameters to be passed}
}
\value{
  list
}
\description{
  Hakee kaikki aanestykset eduskuntarajapinnasta
}
\examples{
# aanestykset <- GetAllAanestykset(...)
}
\author{
  Juuso Haapanen \email{louhos@googlegroups.com}
}
\references{
  citation("sorvi")
}
\keyword{eduskunta}

