\name{GetAllAanestykset}
\alias{GetAllAanestykset}
\title{Hakee eduskunnan äänestysrajanpinnan kautta kaikki aanestykset}
\usage{
  aanestykset <- GetAllAanestykset()
}
\arguments{
}
\value{
  lista kaikista aanestyksista
}
\description{
  Hakee kaikki aanestykset
}
\author{
  Juuso Haapanen
  \email{sorvi-commits@lists.r-forge.r-project.org}
}

