\name{GetEdustajaData}
\alias{GetEdustajaData}
\title{Hakee äänestyksen tiedot äänestystunnisteen perusteella}
\usage{
  aanestykset <- GetAllAanestykset(tunniste)
}
\value{
  Äänestyksen tulos data.framena, jossa kansanedustajan nimi, puolue ja äänestystulos
}
\description{
  Hakee tunnisteella äänestykset edustajaittain.
}
\author{
  Juuso Haapanen
  \email{sorvi-commits@lists.r-forge.r-project.org}
}

