\name{GetEdustajanAanestykset}
\alias{GetEdustajanAanestykset}
\title{Hakee tietyn edustajan kaikki äänestykset}
\usage{
  aanestykset <- GetEdustajanAanestykset('Sukunimi Etunimi')
}
\value{
  Tietyn kansanedustajan kaikki äänestykset
}
\description{
  Hakee kaikki äänestykset tietyltä edustajalta
}
\author{
  Juuso Haapanen
  \email{sorvi-commits@lists.r-forge.r-project.org}
}

