\name{haeHakusanalla}
\alias{haeHakusanalla}
\title{Hae eduskunnan äänestysdataa tietyllä hakusanalla}
\usage{
  aanestykset <- haeHakusanalla('luonto')
}
\value{
  lista äänestyksistä tietyllä hakusanalla
}
\description{
  Hakee tietyllä hakusanalla kaikki äänestykset
}
\author{
  Juuso Haapanen
  \email{sorvi-commits@lists.r-forge.r-project.org}
}

