% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_names.R
\name{harmonize_names}
\alias{harmonize_names}
\title{Harmonize names}
\usage{
harmonize_names(x, synonymes, remove.unknown = FALSE,
  check.synonymes = TRUE, mode = "exact.match", include.lowercase = TRUE,
  verbose = FALSE, from = NULL, to = NULL, ignore_empty = FALSE,
  include.original = FALSE)
}
\arguments{
\item{x}{A character vector}

\item{synonymes}{synonyme table with the fields 'synonyme' and 'name'}

\item{remove.unknown}{Logical. Remove terms that do not have synonymes.}

\item{check.synonymes}{Check the synonyme table}

\item{mode}{'exact.match' replaces the terms based on the synonyme list if an exact match is  found; 'match' replaces the parts that match synonymes; 'recursive' replaces all (sub)strings recursively in the same order as in the synonyme table}

\item{include.lowercase}{Include also lowercase versions of the synonymes. Only works with the exact.match mode}

\item{verbose}{verbose}

\item{from}{String. If given, use the field with this name in the synonymes table as the synonyme list}

\item{to}{String. If given, convert the names corresponding to the 'from' field in this format}

\item{ignore_empty}{Ignore entries with an empty name}

\item{include.original}{Include the original query and result a data frame. Otherwise, return a vector of converted entries.}
}
\value{
Harmonized vector, or a data.frame if include.original is TRUE
}
\description{
Harmonize names
}
\examples{
\dontrun{x2 <- harmonize_names(x, synonymes)}
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("sorvi")
}
\keyword{utilities}

