% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/harmonize_names.R
\name{harmonize_names}
\alias{harmonize_names}
\title{harmonize_names}
\usage{
harmonize_names(x, synonymes, remove.unknown = FALSE,
  check.synonymes = TRUE, mode = "exact.match")
}
\arguments{
\item{x}{A character vector}

\item{synonymes}{synonyme table with the fields 'synonyme' and 'name'}

\item{remove.unknown}{Logical. Remove terms that do not have synonymes.}

\item{check.synonymes}{Check the synonyme table}

\item{mode}{'exact.match' replaces the terms based on the synonyme list if an exact match
       is  found; 'recursive' replaces all (sub)strings recursively in the same order as
       in the synonyme table}
}
\value{
Harmonized vector
}
\description{
Harmonize names
}
\examples{
\dontrun{x2 <- harmonize_names(x, synonymes)}
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("sorvi")
}
\keyword{utilities}

