% Generated by roxygen2 (4.0.1): do not edit by hand
\name{is_url}
\alias{is_url}
\title{Check if the given object is an url string}
\usage{
is_url(s)
}
\arguments{
\item{s}{input object to check

Returns:}
}
\value{
TRUE/FALSE indicating whether the input string is a valid URL.
}
\description{
Arguments:
}
\examples{
is_url("http://aa.px")
}
\author{
Leo Lahti \email{louhos@googlegroups.com}
}
\references{
See citation("sorvi")
}
\keyword{utilities}

