% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_sorvi_data.R
\name{load_sorvi_data}
\alias{load_sorvi_data}
\title{Supporting Data}
\usage{
load_sorvi_data(data.id, verbose = TRUE)
}
\arguments{
\item{data.id}{data ID to download (see details)}

\item{verbose}{verbose}
}
\value{
Data set. The format depends on the data.
}
\description{
Load custom data sets.
}
\details{
The following data sets are available:
   \itemize{
     \item{translation_provinces}{Translation of Finnish province (maakunta) names (Finnish, English).}
   }
}
\examples{
translations <- load_sorvi_data("translation_provinces")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("sorvi")
}
\keyword{utilities}

