% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_sorvi_data.R
\name{load_sorvi_data}
\alias{load_sorvi_data}
\title{Load custom data sets}
\usage{
load_sorvi_data(data.id, verbose = TRUE)
}
\arguments{
\item{data.id}{data ID to download (suffix before .rda). Investigate the contents of the url path to check data.ids}

\item{verbose}{verbose}
}
\value{
translations
}
\description{
Load custom data sets
}
\examples{
# translations <- load_sorvi_data("translations")

}
\references{
See citation("sorvi")
}
\author{
Leo Lahti \email{louhos@googlegroups.com}
}
\keyword{utilities}
