\name{LoadData}
\alias{LoadData}
\title{LoadData}
\usage{
  LoadData(data.id,
    sorvi.data.url = "http://beta.datavaalit.fi/storage/louhos/")
}
\arguments{
  \item{data.id}{}

  \item{sorvi.data.url}{Data URL path}
}
\value{
  url connection
}
\description{
  LoadData
}
\examples{
# LoadData("MML")
}
\author{
  Leo Lahti
  \email{sorvi-commits@lists.r-forge.r-project.org}
}
\references{
  See citation("sorvi")
}
\keyword{utilities}

