% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sorvi-package.R
\docType{package}
\name{sorvi-package}
\alias{sorvi-package}
\alias{sorvi}
\title{Algorithmic Tools for Open Data in Finland}
\description{
The sorvi package hosts various functions that are mainly helpful in rOpenGov 
package maintenance, package authoring and drawing graphs for presentations. 
Additionally it has some functions that do not (yet) have their own package
but are useful in some contexts.
}
\details{
\tabular{ll}{
Package: \tab sorvi\cr
Type: \tab Package\cr
Version: \tab 0.8.21.9000\cr
Date: \tab 2010-2023\cr
License: \tab BSD 2-clause\cr
LazyLoad: \tab yes\cr
}
}
\examples{
library(sorvi)
}
\references{
See citation("sorvi")
\url{https://github.com/rOpenGov/sorvi}
}
\author{
Leo Lahti, Juuso Parkkinen, Jussi Paananen, Joona Lehtomaki, Einari Happonen, Juuso Haapanen, and Pyry Kantanen \email{louhos@googlegroups.com}
}
\keyword{package}
