% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{sorvi-package}
\alias{sorvi}
\alias{sorvi-package}
\title{Algorithmic Tools for Open Data in Finland}
\description{
Brief summary of the sorvi package
}
\details{
\tabular{ll}{
Package: \tab sorvi\cr
Type: \tab Package\cr
Version: \tab See sessionInfo() or DESCRIPTION file\cr
Date: \tab 2010-2014\cr
License: \tab BSD-2-clause\cr
LazyLoad: \tab yes\cr
}

Algorithmic core tools for Open Data in Finland
}
\examples{
library(sorvi)
}
\author{
Leo Lahti, Juuso Parkkinen, Jussi Paananen, Joona Lehtomaki, Einari Happonen, and Juuso Haapanen \email{louhos@googlegroups.com}
}
\references{
See citation("sorvi")
\url{http://louhos.github.com/sorvi}
}
\keyword{package}

