% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{strstrip_single}
\alias{strstrip_single}
\title{Remove spaces from a single string}
\usage{
strstrip_single(s)
}
\arguments{
\item{s}{string}
}
\value{
string without spaces
}
\description{
Remove spaces from a single string
}
\author{
Leo Lahti \email{louhos@googlegroups.com}
}
\references{
See citation("sorvi")
}
\keyword{internal}

