% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hetu.R
\name{valid_hetu}
\alias{valid_hetu}
\title{Validate a Finnish personal identification number (HETU).}
\usage{
valid_hetu(hetu)
}
\arguments{
\item{hetu}{Finnish personal identification number as a character vector, or vector of identification numbers as a character vectors.}
}
\value{
Is the given string a valid Finnish personal identification number, TRUE or FALSE.
}
\description{
Validate a Finnish personal identification number (HETU).
}
\examples{
valid_hetu("010101-0101") # TRUE
valid_hetu("010101-010A") # FALSE
}
\author{
Jussi Paananen \email{louhos@googlegroups.com}
}
\seealso{
\code{\link{hetu}} For extracting information from a Finnish personal identification number.
}

