% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hetu.R
\name{valid_hetu}
\alias{valid_hetu}
\title{Finnish personal identification number validator}
\usage{
valid_hetu(hetu)
}
\arguments{
\item{hetu}{Finnish personal identification number as a character vector, 
or vector of identification numbers as a character vectors.}
}
\value{
Is the given string a valid Finnish personal identification number, 
	   \code{TRUE} or \code{FALSE}.
}
\description{
Validate Finnish personal identification numbers (hetu).
}
\examples{
valid_hetu("010101-0101") # TRUE
valid_hetu("010101-010A") # FALSE
}
\seealso{
\code{\link{hetu}} For extracting information from Finnish 
	    personal identification numbers.
}
\author{
Jussi Paananen \email{louhos@googlegroups.com}
}
